+++ common/models/userModel.js
@@ -39,9 +39,7 @@ let userSchema   = mongoose.Schema({
     verificationText  : String,
     facebookProfileId : String,
     googleProfileId   : String,
-    dropboxProfileId  : String,
     microsoftProfileId: String,
-    twitterUserName   : String
   },
   info        : {
     registrationDate: Date,
@@ -267,41 +265,6 @@ function getGoogleUser(profileId, callback) {
  */
 function getMicrosoftUser(profileId, callback) {
   User.find({'login.microsoftProfileId': profileId}, function (err, docs) {
-    if (err) {
-      return callback(err);
-    }
-    if (docs.length === 0) {
-      return callback();
-    }
-    callback(null, docs[0]);
-  });
-}
-
-
-/**
- * Returns a user by its Dropbox profile
- * @param profileId
- * @param callback
- */
-function getDropboxUser(profileId, callback) {
-  User.find({'login.dropboxProfileId': profileId}, function (err, docs) {
-    if (err) {
-      return callback(err);
-    }
-    if (docs.length === 0) {
-      return callback();
-    }
-    callback(null, docs[0]);
-  });
-}
-
-/**
- * Returns a user by its Twitter profile
- * @param profileId
- * @param callback
- */
-function getTwitterUser(profileId, callback) {
-  User.find({'login.twitterUserName': profileId}, function (err, docs) {
     if (err) {
       return callback(err);
     }
@@ -665,210 +628,6 @@ function findOrCreateMicrosoftUser(profile, callback) {
 }
 
 
-/**
- * Find or create a user logging in with Dropbox
- * @param profile
- * @param callback
- * @returns {*}
- */
-function findOrCreateDropboxUser(profile, callback) {
-  logger.info('findOrCreateDropboxUser', profile);
-  if (!_.isObject(profile) || !_.isString(profile.id)) {
-    return callback(new Error('invalid profile supplied'));
-  }
-
-
-  if (profile.provider !== 'dropbox') {
-    logger.info('This is not a dropbox account: ' + profile.provider);
-    callback(new Error('not a dropbox account: ' + profile.provider));
-  }
-
-  // Try to get the user
-  getDropboxUser(profile.id, function (err, user) {
-    if (err) {
-      return callback(err);
-    }
-    if (!user) {
-      // The user is not here, try to find him with the email-address
-      let emailAddress = _.isArray(profile.emails) ? profile.emails[0].value : undefined;
-
-      function saveNewDropboxUser() {
-        let newUser                    = new User();
-        newUser._id                    = emailAddress || profile.id;
-        newUser.login.dropboxProfileId = profile.id;
-        newUser.info.dropbox           = {
-          emails     : _.get(profile, 'emails', []),
-          name       : _.get(profile, 'name', {}),
-          displayName: _.get(profile, 'displayName', '?')
-        };
-        newUser.info.registrationDate  = new Date();
-        newUser.login.verifiedEmail    = true; // Dropbox does not need verification
-        newUser.personalData.forename  = _.get(profile, 'name.givenName', '');
-        newUser.personalData.surname   = _.get(profile, 'name.familyName', profile.displayName);
-        newUser.personalData.email     = emailAddress ? emailAddress : profile.id; // using profile id as email alternative
-        newUser.personalData.avatar    = _.isArray(profile.photos) ? profile.photos[0].value : undefined;
-        newUser.save(function (err, savedUser) {
-          if (err) {
-            return callback(err);
-          }
-          logger.info('Created dropbox user', savedUser);
-          // Recursive call, now we'll find this user
-          return findOrCreateDropboxUser(profile, callback);
-        });
-      }
-
-      if (emailAddress) {
-        getUserByMailAddress(emailAddress, function (err, user) {
-          if (err) {
-            return callback(err);
-          }
-          if (user) {
-            // Ok, we know this user. Update profile for dropbox access
-            user.info.dropbox           = {
-              emails     : _.get(profile, 'emails', []),
-              name       : _.get(profile, 'name', {}),
-              displayName: _.get(profile, 'displayName', '?')
-            };
-            user.info.registrationDate  = new Date();
-            user.login.verifiedEmail    = true; // Dropbox does not need verification
-            user.personalData.forename  = _.get(profile, 'name.givenName', '');
-            user.personalData.surname   = _.get(profile, 'name.familyName', profile.displayName);
-            user.login.dropboxProfileId = profile.id;
-            user.personalData.avatar    = _.isArray(profile.photos) ? profile.photos[0].value : undefined;
-            user.save(function (err) {
-              if (err) {
-                return callback(err);
-              }
-              logger.info('Upgraded user ' + emailAddress + ' for dropbox access');
-              // Recursive call, now we'll find this user
-              return findOrCreateDropboxUser(profile, callback);
-            });
-            return;
-          }
-
-          // We do not know this user. Add him/her to the list.
-          saveNewDropboxUser();
-        });
-        return;
-      }
-      // No email address (somehow an annonymous drobox user). Add as new User
-      return saveNewDropboxUser();
-    }
-
-    // User found, update
-    user.info.dropbox        = {
-      emails     : _.get(profile, 'emails', []),
-      name       : _.get(profile, 'name', {}),
-      displayName: _.get(profile, 'displayName', '?')
-    };
-    user.personalData.avatar = _.isArray(profile.photos) ? profile.photos[0].value : undefined;
-    updateUser(user, null, callback);
-  });
-}
-
-/**
- * Find or create a user logging in with Twitter. The username is the key for twitter profiles, we don't have
- * an email address!
- * @param profile
- * @param callback
- * @returns {*}
- */
-function findOrCreateTwitterUser(profile, callback) {
-  logger.info('findOrCreateTwitterUser', profile);
-  if (!_.isObject(profile) || !_.isString(profile.username)) {
-    return callback(new Error('invalid profile supplied'));
-  }
-
-  if (profile.provider !== 'twitter') {
-    logger.info('This is not a twitter account: ' + profile.provider);
-    callback(new Error('not a twitter account: ' + profile.provider));
-  }
-
-
-  // This is the twitter user name
-  let username = '@' + profile.username;
-
-
-  // Try to get the user
-  getTwitterUser(username, function (err, user) {
-    if (err) {
-      return callback(err);
-    }
-    if (!user) {
-      // The user is not here, try to find him with the email-address
-      let emailAddress = _.isArray(profile.emails) ? profile.emails[0].value : undefined;
-
-      function saveNewTwitterUser() {
-        let newUser                   = new User();
-        newUser._id                   = emailAddress || username;
-        newUser.login.twitterUserName = username;
-        newUser.info.twitter          = {
-          username   : _.get(profile, 'username', []),
-          displayName: _.get(profile, 'displayName', '?')
-        };
-        newUser.info.registrationDate = new Date();
-        newUser.login.verifiedEmail   = true; // Twitter does not need verification
-        newUser.personalData.forename = _.get(profile, 'name.givenName', '');
-        newUser.personalData.surname  = _.get(profile, 'name.familyName', profile.displayName);
-        newUser.personalData.email    = emailAddress ? emailAddress : username; // using profile id as email alternative
-        newUser.personalData.avatar   = _.isArray(profile.photos) ? profile.photos[0].value : undefined;
-        newUser.save(function (err, savedUser) {
-          if (err) {
-            return callback(err);
-          }
-          logger.info('Created twitter user', savedUser);
-          // Recursive call, now we'll find this user
-          return findOrCreateTwitterUser(profile, callback);
-        });
-      }
-
-      if (username) {
-        getUserByMailAddress(username, function (err, user) {
-          if (err) {
-            return callback(err);
-          }
-          if (user) {
-            // Ok, we know this user. Update profile for twitter access
-            user.info.twitter          = {
-              username   : _.get(profile, 'username', []),
-              displayName: _.get(profile, 'displayName', '?')
-            };
-            user.info.registrationDate = new Date();
-            user.login.verifiedEmail   = true; // Twitter does not need verification
-            user.personalData.forename = _.get(profile, 'name.givenName', '');
-            user.personalData.surname  = _.get(profile, 'name.familyName', profile.displayName);
-            user.login.twitterUserName = username;
-            user.personalData.avatar   = _.isArray(profile.photos) ? profile.photos[0].value : undefined;
-            user.save(function (err) {
-              if (err) {
-                return callback(err);
-              }
-              logger.info('Upgraded user ' + emailAddress + ' for twitter access');
-              // Recursive call, now we'll find this user
-              return findOrCreateTwitterUser(profile, callback);
-            });
-            return;
-          }
-
-          // We do not know this user. Add him/her to the list.
-          saveNewTwitterUser();
-        });
-        return;
-      }
-      // No email address (somehow an annonymous drobox user). Add as new User
-      return saveNewTwitterUser();
-    }
-
-    // User found, update
-    user.info.twitter        = {
-      username   : _.get(profile, 'username', []),
-      displayName: _.get(profile, 'displayName', '?')
-    };
-    user.personalData.avatar = _.isArray(profile.photos) ? profile.photos[0].value : undefined;
-    updateUser(user, null, callback);
-  });
-}
-
 module.exports = {
   Model: User,
 
@@ -882,7 +641,5 @@ module.exports = {
   countUsers               : countUsers,
   findOrCreateFacebookUser : findOrCreateFacebookUser,
   findOrCreateGoogleUser   : findOrCreateGoogleUser,
-  findOrCreateDropboxUser  : findOrCreateDropboxUser,
-  findOrCreateTwitterUser  : findOrCreateTwitterUser,
   findOrCreateMicrosoftUser: findOrCreateMicrosoftUser,
 };