+++ editor/app.js
@@ -73,8 +73,6 @@ let initServer = function () {
   // Define Strategy, login
   passport.use(authStrategy.facebookStrategy);
   passport.use(authStrategy.googleStrategy);
-  passport.use(authStrategy.dropboxStrategy);
-  passport.use(authStrategy.twitterStrategy);
   passport.use(authStrategy.localStrategy);
   passport.use(authStrategy.microsoftStrategy);
   // Session serializing of the user